requirements.txt
streamlit
fal-client
python-dotenv
Pillow
